Moving BaseApp to asyncworker.base
